task 1

+++ 0x0B-malloc_free/1-strdup.c
@@ -21,7 +21,7 @@ char *_strdup(char *str)
 	{
 		return (NULL);
 	}
-	while (i <= strlen(str))
+	while (i <= (int)strlen(str))
 	{
 		myStr[i] = str[i];
 		i++;
